match-service/requirements.txt
Flask==3.1.1
python-dotenv==1.1.0
langchain_core==0.3.60
langchain_mistralai==0.2.10
pydantic==2.11.4
flask-cors